requirements/common.txt
celery[redis]==5.3.6
Flask==3.0.0
Flask-APScheduler==1.13.1
flask-cors==4.0.0
Flask-SQLAlchemy==3.1.1
flower==2.0.1
gevent==23.9.1
gunicorn==21.2.0
librosa==0.10.1
moviepy==1.0.3
numpy==1.26.2
Pillow==10.1.0
pydub==0.25.1
python-dotenv==1.0.0
Requests==2.31.0
tqdm==4.66.1
